Fix some contracts issues
